Consolidate RDF and RDG tests (part 1)
