+++ src/Http/Http.Extensions/gen/RequestDelegateGeneratorSources.cs
@@ -133,6 +133,12 @@ namespace Microsoft.AspNetCore.Http.Generated
                     return (false, default);
                 }
             }
+
+            if (!allowEmpty)
+            {
+                httpContext.Response.StatusCode = StatusCodes.Status400BadRequest;
+            }
+
             return (false, default);
         }
     }
