+++ src/Http/Http.Extensions/test/RequestDelegateGenerator/Baselines/MapAction_MultipleStringParam_StringReturn.generated.txt
@@ -252,6 +252,12 @@ namespace Microsoft.AspNetCore.Http.Generated
                     return (false, default);
                 }
             }
+
+            if (!allowEmpty)
+            {
+                httpContext.Response.StatusCode = StatusCodes.Status400BadRequest;
+            }
+
             return (false, default);
         }
     }
