+++ src/Http/Http.Extensions/test/RequestDelegateGenerator/Baselines/MapAction_SingleComplexTypeParam_StringReturn.generated.txt
@@ -244,6 +244,12 @@ namespace Microsoft.AspNetCore.Http.Generated
                     return (false, default);
                 }
             }
+
+            if (!allowEmpty)
+            {
+                httpContext.Response.StatusCode = StatusCodes.Status400BadRequest;
+            }
+
             return (false, default);
         }
     }
